+++ libpd/pure-data/src/g_vumeter.c
@@ -655,7 +655,7 @@ static void *vu_new(t_symbol *s, int argc, t_atom *argv)
         iemgui_new_getnames(&x->x_gui, 1, argv);
         ldx = (int)atom_getintarg(4, argc, argv);
         ldy = (int)atom_getintarg(5, argc, argv);
-        iem_inttofstyle(&x->x_gui.x_fsf, atom_getintarg(6, argc, argv));
+        iem_inttofstyle(&x->x_gui.x_fsf, (int)atom_getintarg(6, argc, argv));
         fs = (int)atom_getintarg(7, argc, argv);
         bflcol[0] = (int)atom_getintarg(8, argc, argv);
         bflcol[2] = (int)atom_getintarg(9, argc, argv);
@@ -663,7 +663,7 @@ static void *vu_new(t_symbol *s, int argc, t_atom *argv)
     }
     else iemgui_new_getnames(&x->x_gui, 1, 0);
     if((argc == 12)&&IS_A_FLOAT(argv,11))
-        iem_inttosymargs(&x->x_gui.x_isa, atom_getintarg(11, argc, argv));
+        iem_inttosymargs(&x->x_gui.x_isa, (int)atom_getintarg(11, argc, argv));
     x->x_gui.x_draw = (t_iemfunptr)vu_draw;
 
     x->x_gui.x_fsf.x_snd_able = 0;
